+++ store/postgres/src/models.rs
@@ -1,7 +1,7 @@
 use bigdecimal::BigDecimal;
 use diesel::pg::Pg;
 use diesel::serialize::{self, Output, ToSql};
-use diesel::sql_types::{Bool, Float, Integer, Jsonb, Numeric, Text, VarChar};
+use diesel::sql_types::{Bool, Double, Integer, Jsonb, Numeric, Text, VarChar};
 use graph::serde_json;
 use std::io::Write;
 use std::str::FromStr;
@@ -42,10 +42,10 @@ impl ToSql<Bool, Pg> for SqlValue {
     }
 }
 
-impl ToSql<Float, Pg> for SqlValue {
+impl ToSql<Double, Pg> for SqlValue {
     fn to_sql<W: Write>(&self, out: &mut Output<W, Pg>) -> serialize::Result {
         match self.0 {
-            Value::Float(ref f) => <f32 as ToSql<Float, Pg>>::to_sql(&f, out),
+            Value::Float(ref f) => <f64 as ToSql<Double, Pg>>::to_sql(&f, out),
             _ => panic!("Failed to convert non-float attribute value to float in SQL"),
         }
     }
